ref: calling user location function on the client side
